Unconditionally refresh on config-changed.
This hook will fire when config is changed, but also on upgrade-charm.
Because the hook and not a reactive flag is used, it is OK to
unconditionally refresh the service whether the config actually changed
or not. This ensures upgrade-charm can apply changes to the config that
might be in template or other charm updates.